chore: internationalize genres
* Internationalized genres.

* chore: internationalize game detail title


Approved-by: frupnik
Approved-by: Mauro Leandro Baez

+++ models/src/main/java/ar/edu/itba/paw/dtos/ReviewFilter.java
@@ -3,12 +3,12 @@ package ar.edu.itba.paw.dtos;
 import java.util.List;
 
 public class ReviewFilter {
-    private final List<Long> gameGenresFilter;
-    private final List<Long> reviewerPreferencesFilter;
+    private final List<Integer> gameGenresFilter;
+    private final List<Integer> reviewerPreferencesFilter;
     private final ReviewOrderCriteria reviewOrderCriteria;
     private final OrderDirection orderDirection;
 
-    public ReviewFilter(List<Long> gameGenresFilter, List<Long> reviewerPreferencesFilter, ReviewOrderCriteria reviewOrderCriteria, OrderDirection orderDirection) {
+    public ReviewFilter(List<Integer> gameGenresFilter, List<Integer> reviewerPreferencesFilter, ReviewOrderCriteria reviewOrderCriteria, OrderDirection orderDirection) {
         this.gameGenresFilter = gameGenresFilter;
         this.reviewerPreferencesFilter = reviewerPreferencesFilter;
         this.reviewOrderCriteria = reviewOrderCriteria;
@@ -19,11 +19,11 @@ public class ReviewFilter {
         this(null, null, ReviewOrderCriteria.REVIEW_DATE, OrderDirection.DESCENDING);
     }
 
-    public List<Long> getGameGenresFilter() {
+    public List<Integer> getGameGenresFilter() {
         return gameGenresFilter;
     }
 
-    public List<Long> getReviewerPreferencesFilter() {
+    public List<Integer> getReviewerPreferencesFilter() {
         return reviewerPreferencesFilter;
     }
